requirements.txt
azure-core==1.27.1
azure-devops==7.1.0b3
colorama==0.4.6
halo==0.0.31
GitPython==3.1.31
msrest==0.7.1
tabulate==0.9.0
python-dotenv==1.0.0